+++ Weiche/Platforms.cs
@@ -14,7 +14,7 @@ namespace Weiche
             {
                 string name;
                 double radius = Weichengenerator.radius;
-                int segmente = Weichengenerator.segmente;
+                double segmente = Weichengenerator.segmente;
                 double gaugeoffset = Weichengenerator.gaugeoffset;
                 string launchpath = Weichengenerator.path;
                 string texture_format = Weichengenerator.texture_format;
